Move HeadwordEntry to Entity package as it is an entity
