+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/mixin/identifiable/entity/HeadwordEntryMixIn.java
@@ -1,10 +1,9 @@
-package de.digitalcollections.model.jackson.mixin.semantic;
+package de.digitalcollections.model.jackson.mixin.identifiable.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import de.digitalcollections.model.jackson.mixin.identifiable.entity.EntityMixIn;
-import de.digitalcollections.model.semantic.HeadwordEntry;
+import de.digitalcollections.model.identifiable.entity.HeadwordEntry;
 
 @JsonDeserialize(as = HeadwordEntry.class)
 @JsonIgnoreProperties(ignoreUnknown = true)
