+++ dc-model-jackson/src/test/resources/serializedTestObjects/identifiable/entity/HeadwordEntry.json
